+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-- Add `description` data to `myRegistrations` single view (#5159)
+- Add more data to the "my registrations" single view (#5159, #5172)
 - Add `Event.timeSlots` (#5149)
 - Add a `TimeSlot` model and repository (#5039)
 - Add more columns to the FE list views (#5119, #5131, #5150, #5152)